#### What this PR does
Yesterday we merged [a PR](https://github.com/grafana/mimir/pull/5668) to improve the active series tracker benchmark. Then today, while reviewing https://github.com/grafana/mimir/pull/5678, I realised we don't test the concurrency with `Purge()` too (which takes an exclusive lock on the stripe). In this PR I'm improving the benchmark again, to also stress `UpdateSeries()` while `Purge()` is being called.

To make the `Purge()` effective (that is delete all series each time), I've done few changes to the test:
1. Set active timeout = 0
2. Guarantee that the "current time" passed to `UpdateSeries()` and `Purge()` is always increased. When I pass it to `Purge()` I jump the mocked "current time" to the future so that we're sure it's well above the last calls to `UpdateSeries()`

I've run the updated benchmark. "Before" is with 512 stripes, "After" is with 128 stripes. Looking at absolute numbers, I can't see any significant difference:

```
goos: darwin
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester/activeseries
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                                                                            │  before.txt  │              after.txt              │
                                                                                            │    sec/op    │    sec/op     vs base               │
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_50,_purge_=_false-12           68.14n ± ∞ ¹   75.07n ± ∞ ¹  +10.17% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_50,_purge_=_true-12            64.86n ± ∞ ¹   79.56n ± ∞ ¹  +22.66% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_100,_purge_=_false-12          86.32n ± ∞ ¹   92.71n ± ∞ ¹   +7.40% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_100,_purge_=_true-12           79.76n ± ∞ ¹   89.92n ± ∞ ¹  +12.74% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_500,_purge_=_false-12          93.34n ± ∞ ¹   97.75n ± ∞ ¹        ~ (p=0.151 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_500,_purge_=_true-12           88.49n ± ∞ ¹   90.20n ± ∞ ¹   +1.93% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_1000,_purge_=_false-12         100.8n ± ∞ ¹   101.5n ± ∞ ¹        ~ (p=0.119 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_1000,_purge_=_true-12          90.65n ± ∞ ¹   91.17n ± ∞ ¹        ~ (p=0.841 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_50,_purge_=_false-12     52.07n ± ∞ ¹   62.66n ± ∞ ¹  +20.34% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_50,_purge_=_true-12      51.87n ± ∞ ¹   67.05n ± ∞ ¹  +29.27% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_100,_purge_=_false-12    52.22n ± ∞ ¹   56.10n ± ∞ ¹   +7.43% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_100,_purge_=_true-12     49.77n ± ∞ ¹   55.29n ± ∞ ¹  +11.09% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_500,_purge_=_false-12    49.68n ± ∞ ¹   55.33n ± ∞ ¹  +11.37% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_500,_purge_=_true-12     49.25n ± ∞ ¹   52.67n ± ∞ ¹        ~ (p=0.056 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_1000,_purge_=_false-12   48.95n ± ∞ ¹   49.68n ± ∞ ¹        ~ (p=0.310 n=5)
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_1000,_purge_=_true-12    48.95n ± ∞ ¹   50.06n ± ∞ ¹        ~ (p=0.095 n=5)
geomean                                                                                       64.71n         70.63n         +9.15%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                                            │ before.txt  │              after.txt              │
                                                                                            │    B/op     │    B/op      vs base                │
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_50,_purge_=_false-12           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_50,_purge_=_true-12            0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_100,_purge_=_false-12          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_100,_purge_=_true-12           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_500,_purge_=_false-12          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_500,_purge_=_true-12           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_1000,_purge_=_false-12         0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_1000,_purge_=_true-12          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_50,_purge_=_false-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_50,_purge_=_true-12      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_100,_purge_=_false-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_100,_purge_=_true-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_500,_purge_=_false-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_500,_purge_=_true-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_1000,_purge_=_false-12   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_1000,_purge_=_true-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                                                                 ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                                                                                            │ before.txt  │              after.txt              │
                                                                                            │  allocs/op  │  allocs/op   vs base                │
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_50,_purge_=_false-12           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_50,_purge_=_true-12            0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_100,_purge_=_false-12          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_100,_purge_=_true-12           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_500,_purge_=_false-12          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_500,_purge_=_true-12           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_1000,_purge_=_false-12         0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1,_concurrency_=_1000,_purge_=_true-12          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_50,_purge_=_false-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_50,_purge_=_true-12      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_100,_purge_=_false-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_100,_purge_=_true-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_500,_purge_=_false-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_500,_purge_=_true-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_1000,_purge_=_false-12   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/series_=_1000000,_concurrency_=_1000,_purge_=_true-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                                                                 ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean
```

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`